Merged PR 1724167: Temporarily suppress fast fail from 17130145 to prevent issues when ACI RI's
There's a real issue here, but I have to add another set of asserts at other points to actually catch where we went wrong.

To prevent this specific stack from becoming very noisy as we RI ACI up to main, I'm suppressing it for when we RI.

I'll send another pull to restore it and add more FAST FAIL asserts to catch the real issue after this one is completed and merged.

Related work items: #17130145