Merged PR 1768108: Make Search process based on abstraction, not embedded text buffer details
This rewrites `Search` into an object. You create the object and then `.FindNext()` much like a `Regex` object pattern. Then you can `.Select()` or `.Color()` it to perform the two verbs without actually needing to know the intricate details of where the found item was stored.

The initial performance of this was pretty bad due to `std::vector` creation and copying, so also included is `ScreenInfoTextIterator`, a custom iterator that hides the intricate inner details of the buffer providing only the ability to walk around from cell to cell and retrieve the `wchar_t` data by a simple dereference.

I've added new tests to cover `ScreenInfoTextIterator` and the actual searching logic of `Search`. The color/selection parts are uncovered as they require expensive UIA test work. I've manually verified that they work for both the Find dialog box as before and for the `EnableColorSelection` registry key toggle that makes Ctrl+Shift+# and Alt+Shift+# do a find-and-color in the buffer.

Additionally a side effect of this is that the new search algorithm is capable of finding across lines. That is, a word spilling off the right edge of the buffer and onto the next line will be found correctly. This resolves ridiculously long standing backlog item #614365.

Finally, there's a tiny bonus fix for a `ROW` initialization bug I found with one of the tests.

Related work items: #16115514, #614365