Merged PR 1811180: Merged PR 1794053: Remove assorted types of asserts. Either fast fail or don't
Merged PR 1794053: Remove assorted types of asserts. Either fast fail or don't bother

Since we don't really do chk and have a ton of different types of asserts, we're going to reduce this all down to one pattern.

Either it's important enough to check and fast fail on to generate a Watson bucket all the time, or it should be recovered and keep going.

Many of the asserts we have are indicative of programming failures on our part, so they should crash when we are building/testing them. Others are indicative of stuff that should "never happen". So either they never happen and we won't crash, or they do happen, will crash, and we can identify what sorts of reliability enhancements to make in these areas by their Watson hit level.

I've removed assert(), ASSERT(), NT_ASSERT(), Assert(), and ASSERT_FRE() from the code and replaced things with either a FAIL_FAST_* wil macro or removed it, especially if it's a thing we've never really utilized.

Related work items: #16669592

Related work items: #16669592