Merged PR 1863008: Make a reasonable default assumption if the client area gets too small.
Make a reasonable default assumption if the client area gets too small. For some reason, AEAUTO tests are running with an absolutely TINY window (21 pixels tall) at -32000,-32000. This is causing an assertion (which was CHK only and is now FRE) to fail.

As such, I added a code snippet that instead makes a reasonable assumption that if your window is absolutely tiny and in the middle of nowhere, a 1x1 viewport is good enough for you. I chose 1x1 because I'm sure 0x0 would cascade into a MulDiv or Divide error somewhere else.

Cherry picked from !1856999

Related work items: #17499419