Merged PR 1891123: Fail fast was invalidly dereferencing 0th element
Fail fast was invalidly dereferencing 0th element when the size of the outer vector was 0 for multiplication in the assert. I changed this to instead use a try/catch gsl::span which should suffice for any boundary issues that the FAIL_FAST was attempting to check before copying. #17658035

Related work items: #17658035