Merged PR 1902230: Fix unallocated memory being copyed/moved into row
Remove explicit moves and swaps and copy operators as we're doing nothing special and one of them has a bug swapping/using unallocated memory. Also clean up and equals and remove some of the debug things from ATTR_ROW that are wasting time while we're in here. #17654369

Related work items: #17654369