Merged PR 1968403: Merged PR 1952485: Add a tracelogger to the VT renderer to verify we're doing
Merged PR 1952485: Add a tracelogger to the VT renderer to verify we're doing it right

Add a tracelogger to the VT renderer to verify we're doing it right.

Attached bug is only related, this won't fix that, but this does make it a LOT easier to debug.

Related work items: #17815313

Related work items: #17815313