Merged PR 1965646: Fix bugs that were introduced by incorrect conversion to stl
In !1850186 a lot of changes were made to convert to using stl types. Several of these caused significant bugs with cmd history.

One error was a bug from converting if(memcmp) into if(std::equal) in the code that adds a command to the history. This meant that we reset the history stack when we did not change the command (and failed to reset it if we changed the command but it was the same length).

Another error was a bug where a reverse iteration of commands in FindMatchingCommand was turned into a forward range based iteration. The index used for the result was left iterating in the reverse order, so we were comparing different commands to the result that was being reported.

Another error was a bug where the popup was not being told the current command. This meant the popup did not have the correct command selected and it was possible to walk off the end of the list of commands causing memory corruption and crashes.

These combined have the effect of breaking command history behavior in a bunch of non-obvious ways that were easy to run into.

Related work items: #17774750