Merged PR 1980161: Clear the screen when we don't inherit the cursor
Clear the screen when we don't inherit the cursor. I removed this behavior originally to support inheriting the cursor on WSL, but now for openSSH we *do* need to clear the screen, because they won't be able to inherit the cursor. So now we'll clear the screen _unless_ someone asked for the cursor position.

This also correctly moves the viewport on the terminal down when you issue a `cls`, instead of just erasing the contents of the current viewport.

Cherry picked from !1969471

Related work items: #17263834, #17815688