Merged PR 1979335: A great idea would be to mark pending reads as completed when they're done.
This manifests itself in several ways:
- When running batch scripts with a prompt, cancelling the prompt will result in the prompt line being printed over and over forever.
- When running net use and getting prompted for a password, you can't hit enter to send the password through.

The issue is that when I encapsulated some of the parameters for cooked reads and the pending input data that is carried between requests, it looks like I lost track of clearing the pending state when the read was consumed or otherwise needed to be discarded.

I've added a CompletePending() method that sets the flags as expected and clears out the internal state.

I validated manually on RS_ONECORE_DEP_ACI from 6-15 that both of the above scenarios are broken with the in-box conhost.exe and are resolved when I use an OpenConsole.exe generated from this branch.

Related work items: #17881611