Merged PR 1994977: Fix watson crash when history lists are shuffled for LRU and wait/popup references become invalidated.
Don't use deque, use list. We rearrange the elements in here to maintain an LRU order. Those sorts of 'iterator pointers' do not invalidate when shuffling around a list with .erase, but they do get screwed up with a deque. This is the simplest solution. #17782993

Cherry picked from !1990552

Related work items: #17782993