Merged PR 2007778: pack CharRowCells together to eliminate dead memory
CharRowCell is 3 bytes big but by default 4 bytes are allocated for it for word alignment purposes. this leads to ~100MB being wasted (1 byte at a time) in scenarios where the screen buffer is set very large. we now force the compiler to pack CharRowCell by bytes to eliminate this wasted memory.

Related work items: #17954106