Merged PR 2014192: Remove unproductive fail fast around reading buffer rectangle
This particular fail fast goes off in a race around resizing scenarios where it isn't really important. It is OK for the row line to not perfectly match the size of what is requested because the .resize method will pad it out with appropriate characters if it is too small or trim it down if it is too big. As of now, the fail fast only shows that "something happened" or "is happening" and provides us no value with actually tracing it down and correcting whatever it might be.

I debugged a whole bunch of customer dumps for this and engaged with a few internal customers who were running into it. None of them seemed to be able to reliably repro it or were perturbed by the behavior of their console otherwise. The Time Travel Traces I was able to obtain showed that nothing was really "wrong" here, other than that this fail fast was firing while the console was still in a state where it was "stabilizing" itself from the resize.

If this is still causing problems for end-users at this point, I expect they will report the issues through repro steps and a specific scenario that might show us what is actually happening for them. Until then, we should turn this off to stop the Watson crashes and noise.

Related work items: #17776843