Merged PR 2017864: Fix byte counting problem with pending input.
Fix byte counting problem with pending input. #17987369

This was introduced with the encapsulation of part of the input buffering and pending input stuff. I used string_view and then had a byte counting problem.

Really this whole method should be upended, but for 19H1 hopefully in #18047766

Cherry picked from !2014058

Related work items: #17987369