Merged PR 2043981: Improve Vim exit performance with attr inserts and avoiding vectors
Added two more "special cases" to the head of InsertAttrRuns to improve insertion performance:
If the inserted runs cover the entire length of the would-be row, just straight up use the new run information to cover up the old run information and don't process further.
If we have a run with two colors exactly in it and we're walking down it 1-by-1 to change it from the "right" color to "left" color by inserting a 1-length run with the left side color, identify that scenario and just ++ and -- the two pieces we have instead of doing the whole HeapAlloc business below. (This was exacerbated with some of the UTF-8 work because now we commit a character and attr in the same loop iteration instead of bulk-inserting all characters then bulk-inserting all attrs.)
Fixed major performance drag in ReadConsoleOutput attributed to heap allocation and freeing of piles of vectors for OutputCells.
To resolve this, I made another case of the ScreenInfoTextIterator like was used for search and called it ScreenInfoCellIterator where it can walk the internals in a high performance read-only way and just dereferences into the CHAR_INFO that the API wants without any other kerfluffle. Since these are so similar, I made one inherit from the other and I templatized the tests to cover both for most of the operations.

Related work items: #17779081