Merged PR 2045771: trigger a render update when we change box drawing settings
trigger a render update when we change box drawing settings. there's a race condition where we may not get the message to change the box drawing settings before text is written and drawn the first time, so force an update to make sure that what's shown on the screen is accurate.

Related work items: #17363227