Merged PR 2077044: Undo the wrapping change from before
Don't do anything clever to try and handle wrapped lines. The change wasn't working as expected, and it was causing us to permanently corrupt the contents of the terminal.

Also don't clear the screen with a 2J on resizes, only repaint the buffer contents. This was causing some duplicated output in the terminal

Cherry picked from !2051644

Related work items: #18123777, #18195169