Merged PR 2089354: Don't load user settings in pcon mode
Don't load user settings in pcon mode. This will make the terminal and host more consistent.

Say your defaults were set to black on white, but the terminal is white on black. The terminal would probably get strange vt that's accurately describing the state of the host, but would look weird in the terminal.

This will make the experience consistent, as we're only ever using the hard-coded defaults, and anything from the startupinfo (which the terminal might have requested, so it wants that behavior)

Cherry picked from !2081685

Related work items: #18194695