Merged PR 2102869: Clear internal cache when resetting the rest of the IME data
Internal string-type storage isn't being cleared when the rest of IME data is being cleared. #18258502

This got missed when I detangled the IME from using its weird byte packets for transmission. The strings are getting cached, but not cleared, when the rest of the screen buffer data is being cleared.

This used to go away when the packet was deleted, but that doesn't happen anymore.

Cherry picked from !2098413

Related work items: #18258502