Merged PR 2112822: output from apt install should appear in the scrollback
Currently, we discard output from applications that set the scroll margins to (top, anything) and emit a lot of text in the margins. Case in point: `apt install <anything>`.

We should instead be moving lines that are scrolled off the top of the buffer when the top margin is the top of the viewport into our "scrollback". We accomplish this by shifting the contents of the viewport down by the number of lines scrolled, then scrolling the margins content up by that number of lines, effectively inserting the correct number of lines at the bottom margin.

Cherry picked from !2107686

Related work items: #15019524