Merged PR 2115138: Merged PR 2112865: Better emulate terminal behavior
Merged PR 2112865: Better emulate terminal behavior

Introduce a "virtual bottom" to the buffer. When the viewport moves for any reason other than the user scrolling the window, we'll update the virtual bottom of the buffer to the new viewport's bottom.

In VT mode, whenever we recieve a character, before we process the string with the StateMachine, we'll reset the viewport to the last virtual location it was. Before this change, if you scrolled around in the buffer, and a client app emitted a move cursor sequence, it would move the cursor relative to the scrolled position, not where the viewport actually was when the app last emitted text. For something like ssh, there was no way for it to cause the viewport to "snap back" to where it was before if the user had scrolled it, which could lead to some pretty badly corrupted viewports.

Additionally, if the viewport moves below it's last virtual location, we'll re-initalize those new rows with the current attributes in VT mode. This will emulate the behavior of adding a "new" row to the bottom of the viewport.

Related work items: #15683949, #17415310

Related work items: #15683949, #17415310