Merged PR 2191425: Improve performance of writing console text through avoiding copies with views and custom iterators
The WriteConsole/WriteLine codepath was running completely through OutputCells which make copies and many vectors in order to operate.

For some simpler scenarios that don't actually require a copy of the data, it is more performant to store read-only views and iterate over the already-in-memory data in a structured fashion on inserting each individual cell into the backing buffer instead of processing each level in bulk and allocating another heap location.

This applies the `OutputCellView` and `OutputCellIterator` pattern to the very common simple-text-writing route.

Other areas are currently out of scope and still use `OutputCell`. We can migrate later as need arises or as a part of a future task.

Related work items: #18565221