Merged PR 2224021: Add deferred presentation phase to renderer thread, have DX renderer use it
This adjusts the renderer thread loop to move the lock inside the frame painting and have it unlock during the "presentation" phase of the renderer so the other threads (particularly the IO thread) can keep servicing while the frame is moving from the backing buffer to the screen.

I've activated support for this feature in the DX renderer to increase performance.

Cherry picked from !2204716

Related work items: #18587890