Merged PR 2325969: fix unsigned integer underflow in UiaTextRange scrolling
fix unsigned integer underflow in UiaTextRange scrolling. this will stop the invalid state that one of our asserts caught.

The invalid state is triggered when we try to scroll the viewport and told not to align the text range to the top of the viewport and the end of the range is less than a viewport's amount away from the top of the screen buffer. the fix is to check this state directly instead of performing an unnecessary subtraction first.

Cherry picked from !2314321

Related work items: #18920370