Merged PR 2426455: Resolve invisible cursor and cursor blink rate issues
When I had tried to fix "cursor turds" in scrollback output, I introduced this issue.

As a result, I took a different approach for removing the cursor from the scrolled region.

Instead of removing the cursor from every frame and assuming it would be repainted (which it isn't necessarily anymore...),
I only remove the cursor from the frame during the scroll operations.

This also allowed me to eliminate one of the phases of the renderer base engine that was no longer in use. I had a different idea for it back when I originally wrote it and we've evolved the rendering pipeline differently now especially with this change.

Also included is a little debugging helper that creates a secondary window when set under the debugger to show some of the operations occurring on the backing buffer while a frame is being painted to help visualize what is happening. We can evolve this further over time, if necessary. This code should all be ifdef'd out for release.

Cherry picked from !2421365

Related work items: #17369304, #19163282, #19205344