Merged PR 2439882: Remove commented out fast fails for scroll calculations.
Remove commented out fast fails for scroll calculations.

I looked at this through a handful of scenarios that caused this to fail.

If the font size or the buffer is too small, having it fast fail here isn't really useful because it's too late. The problem has already occurred a long time ago. So getting a crash isn't useful.

If the client area is too small, as was what happened when the test system originally filed issues against these "asserts", then the "reasonable default" I was going to program in was to ask for a scroll bar (because the client area is too small to show the text... it is also too small to show the scroll bar so it doesn't really matter...) Turns out, that's already what's happening here with the commented out fail fasts.

So we may as well just dump them out entirely and solve other problems as they arise.

Related work items: #17663344