Merged PR 2442871: Fix SFC command cut-off by giving proper column count to DBCS function
This resolves the issue with DBCS characters overstriking each other and leaving spaces as seen with the SFC /scannow command on Japanese systems.

The issue stems from the cleanup function that is used to ensure that extraneous trailing pieces are not left behind in the buffer when part of a line is written.

When some of the rest of the writing functions were refactored to stop double-representing the characters as they were being written, that also adjusted the counts for the helper functions that removed extra characters.

This results in the `CleanupDbcsEdgesForWrite` function being too aggressive and accidentally clearing stuff that we meant to write because it had the incorrect length count.

The fix provided is a small patch to resolve the issue down the one route that is affected.

The underlying issue here is that the assorted callers for writing into the buffer are responsible for understanding how "bisects" could occur. A "bisect" is defined as an improper splitting of a leading and trailing character.

I would like to move this logic down down down to where characters are actually placed into the buffer. But there are currently a few assorted ways that this happens. Instead of making this too complicated right now in the mission of fixing the bug, I'll just apply this small fix for now and hopefully come around later and move the logic down lower with the UTF-8 deliverable.

Also, a tiny patch is easily cherry-pickable if we need to backport.

Related work items: #18335873