Merged PR 2442966: Don't always MoveToBottom
This fixes VsCode, Winpty, as well as addresses a UX regression in cmd.exe

Prior to this fix, we'd move the viewport to the last logical position it was in (before the user scrolled it) before any text was parsed by the parser. This would mean that any apps that emitted text in VT mode would act like terminals, and for whatever reason, this was really messing up winpty.

The solution is to only reset the viewport when a VT sequence actually needs to know something about the vertical position of the viewport.

Related work items: #19103917, #19205200