Merged PR 2466178: Make DxEngine understand High DPI
Basically implements the GetProposedFont and UpdateDpi routes in a similar fashion to the GdiEngine.

The DX renderer now handles DPI scaling by basically upscaling the font. It's not super helpful to use the built-in DirectX ways of scaling for DPI given we already have to micromanage the pixel-to-cell count and the pixel client area of the window to get our appropriate layout.

Related work items: #19317306