Merged PR 2481863: Improve performance of `ScrollConsoleScreenBuffer`
OK First, my apologies for how many files this touches.

The primary goal here was to try to increase the performance of `ScrollConsoleScreenBuffer` API.

I did two phases on this:
1. I profiled the API as it was and worked to cut down the amount of time it took. This primarily happened around eliminating excess heap allocations that were occurring for every cell in the buffer. I did this through using our Iterators and Views to get at the underlying data and by creating the `OutputCellRect` class which represents one big block of data that can be filled up in bulk with a single allocation instead of a ton of tiny allocations.
2. I edited the API itself to detect the scenario where complete rows of data are being scrolled. In this case, it now contacts the underlying `TextBuffer` and asks it to just reshuffle the entire rows around inside the container instead of copying/allocating/moving anything.

You can see more intricate details on the GitHub post: https://github.com/Microsoft/console/issues/279

Secondly, I cleaned up some of the helpers here.
As I was working on the iterators, I needed to move around inside the screen buffer or viewport more efficiently. So I moved a bunch of the utility methods from the generic utility class onto `Viewport`, changed all references in the codebase to use the new ones, and added a suite of tests to Viewport.

Related work items: #18736783, #19270439