Merged PR 2530481: Standardize conversion between W/A around string based functions
This gets rid of all usages of the multi-parameter buffer/size versions of the W/A conversion functions from convert.cpp.
I've also gone through and changed the function signatures/handling of all the related API call functions (so to not have to marshal things in and out of views and raw buffers over and over).
I dropped a ton of Hungarian and out-of-date SAL while I was at it as well.

Cherry-picked from !2528442

Related work items: #19430924