Merged PR 2569460: Adds test for cell bisecting and resolves behavior
This adds new tests around the bisecting behavior of characters in the buffer.

There is an intentional change to behavior for the W route where we're trying to insert a full width character at the end of the run, which doesn't make sense and was compensate for elsewhere all over the codebase. I removed one such instance in the bisect clipboard function. Our overall goal is to make sure that the data inside `TextBuffer` is always valid and makes sense.

The A versions of the behavior are maintained for now.

There's additionally a fix to TextAttribute (color data) so it strips out and won't store spacing data by accident (DbcsAttribute).

Cherry picked from !2564211

Related work items: #19430925