Merged PR 2587333: use correct OutputCellIterator constructor for drawing CommandListPopup colors
`OutputCellIterator` has a constructor that takes a `wchar_t` and a constructor that takes a `TextAttribute`. We were passing a `WORD` which  got cast to a `wchar_t` when we really wanted it to be used to create a `TextAttribute`. This PR makes it very explicit which constructor we want in CommandListPopup.

Cherry picked from !2577736

Related work items: #19600567