Merged PR 2587058: VT output machine needs to use attached screen buffer, not just the active one
When screen buffers are created, an output VT engine is added to each one. However, the `IIoProvider` attached to the engine was looking up the globally active screen buffer no matter what to write output data. This means that writing to an inactive, non-displayed screen buffer with VT mode on actually was showing up on the active screen buffer instead.

To resolve this, I made `SCREEN_INFORMATION` also an implementer of `IIoProvider`. This way Virtual Terminal modules can be instantiated either to care about their specific IO objects (like for the output engine case) or instantiated to not care and just always write to the active one (all the other cases I could find, but primary input engine cases.)

To allow `SCREEN_INFORMATION` to inherit/implement anything, I had to adjust the `ConsoleObjectHeader` stuff to no longer be dependent on being the first item in the console objects. This is because adding an interface adds a vtable which was offsetting the `this` pointer and making a mess. This also affected `InputBuffer` as the other consumer of `ConsoleObjectHeader`.

Finally, I've adapted the sample code that was given to us on GitHub for reproduction of this bug into a feature test to ensure that things will continue to work properly in the future for both VT and non-VT output writing modes.

Related work items: #19317384, #9358743