Merged PR 2603292: Restore word-by-word selection behavior for keyboard-based navigation
When converting to use the iterators, I apparently accidentally dropped the check of the return values from the increment/decrement commands that was used to know when we had or had not run into the end of the buffers.

This made us think that we always hit the end of the buffer right away (since the default value is false) causing word-by-word selection to terminate very quickly.

I've restored the boolean assignment with the return value to fix this. (I can't make it [[nodiscard]] because there are other valid users of the utility function that truly do not care whether it was true or false.)

I've also added some very basic tests to word by word selection. They're not comprehensive by any means, but they should prevent this specific sort of thing from happening in the future. Note they avoid most of the utilities and shorthand used by the real `WordByWordSelection` method so they aren't following the same codepaths to validate the behavior.

Cherry picked from !2603217

Related work items: #19629061