Merged PR 2603921: Restore spaces parameter of WriteCharLegacy to returning cell count (not character count)
When the buffer was being converted to use the new writing methods, there was an oversight that turned the `pcSpaces` parameter of `WriteCharsLegacy` from returning cell count to returning just character count.

It turns out that `pcSpaces` major functionality is for `COOKED_READ_DATA` and the `cmdline.cpp` file to know how many columns of the buffer they're covering up so they can be erased with "spaces" later (thus the name). This means column count because spaces are halfwidth.

The fix here is to use the new `OutputCellIterator` functionality that reports/abstracts the cell difference before and after write to use as the incrementor of the pointer. Then the correct number of spaces will be stored in the `COOKED_READ_DATA` to be erased later.

Also included is an addition to the command line tests that covers this specific scenario by injecting some Japanese text into the buffer via the input routes, ensuring the correct text is in there, then calling the function off of `CommandLine` class that is triggered normally when the user presses Ctrl+Home.

Related work items: #19652455