Merged PR 2626625: Fix the alt buffer
See OpenConsole !2587058

When you switch to the alt buffer, we set the alt buffer as gci's active buffer (`gci.pCurrentScreenBuffer`).
gci's implementation of `IIoProvider` always returned `pCurrentScreenBuffer`.
we added that interface to ScreenInfo, but implemented it as always returning `this`, instead of the alt buffer, if the alt buffer is attached.
So when we'd switch into the alt buffer, instead of VT calls getting sent to the alt buffer, they'd get sent to the main buffer.

Related work items: #19734835