Merged PR 2643670: Restore clipping, boundary conditions, and negative processing to WriteConsoleOutput API
When refactoring the `WriteConsoleOutput` API in both RS5 and 19H1, we've lost some of the assorted boundary conditions that were possible in the legacy console mode.

This PR writes tests for those boundary conditions exercised against the v1 legacy console and then adds the restrictions to the API servicing call inside the current version of the console host.

This happens to also resolve an issue with `rclone`.

I also added a quick way to flip the feature tests to run against the v1 console for future development of tests.

Related work items: #19313218, #19549237