Merged PR 2718791: Buffer writes per frame to minimize kernel WriteFile calls slowing down ConPTY
Buffer writes per frame to minimize kernel WriteFile calls slowing down ConPTY output. #19916177

We just use a simple std::string as the buffer and append during the calling to Write while making the frame and emit the entire string (and clear it) to WriteFile with a single kernel call on EndPaint.

Related work items: #19916177