Merged PR 2770483: Always clear the resize supression flag
In the VT renderer, we'd suppress the first resize message, because it'd usually be redundant. However, at some point, I made it so only resizes that actually changed the viewport size got renderered out. During that change, I overlooked the fact that the first resize is always the same size as the current viewport size, and we'd only clear the suppression flag when a resize with different dimensions came through.
So, always clear the suppression flag.
fixes #19408543.
Adds a test too.

Cherry picked from !2770188

Related work items: #19408543