Merged PR 2963982: Adapt email about setting up TESTPASSES into the UniversalTest document
Adapt email about setting up TESTPASSES into the UniversalTest document to provide some record and background on how these things are put together. #20623477

I wrote an email to the ES Universal Test team while I was migrating us to the new Source Is Truth initiative describing, in extreme detail, how we connect everything up to run console's tests with the new system. I've adapted it here and am committing it into the repository for future reference.

Try not to be too nitpickey about this because I don't want to let the perfect documentation be the enemy of going from 0 documentation on this subject to some documentation and I want us all to feel encouraged going forward to just start writing stuff down even if it's incomplete or not perfect.

Related work items: #20623477