Merged PR 2994529: Attempt to adjust TestReadFileEcho to improve lab reliability and debuggability
There's a sporatic failure in this test in the lab. I cannot reproduce it locally nor in TDP/Nebula even with a loop run (to run it a ton of times in a row).

As such, I'm going to take two guesses here. It looks like the async is getting screwed up so I've changed it to be cancellable and added a cleanup step to try to resolve that. I've also added the parameter to the test job to make it screen capture the failure so I can see what's going wrong.

If neither of these end up working, I'll disable this test for the RI-TP lab and keep it enabled for CI and desktop inner loop runs.

I've also documented how to do some of this reproing in an MD file.

Related work items: #20742113