Merged PR 3014711: Merged PR 3013827: [Terminal] Fix Scrolling in Release builds
Merged PR 3013827: [Terminal] Fix Scrolling in Release builds

The terminal-initiated scrolling events would come out of order, or during scrolling in Release builds, and that would cause the viewport to scroll to the incorrect position.

This changes the way terminal-initiated scrolling works. Now, when we send a scroll event, we'll stash that requested position. When the scroll event is handled, we'll inspect that last requested position. If it's set, and we're not handling the message that last set that position, we'll just ignore the message. Mouse scrolling, or other user-initiated scrolling will clear that out.

Related work items: #20642296, #20752238

Related work items: #20642296, #20752238