Write profiles.json file to RoamingAppData folder in unpackaged scenario
Currently, in the scenario where Cascadia is run outside of the package, profiles.json is written to the current location of the executable. This change modifies the save location to be under %userprofile%\appdata\Roaming\Microsoft\Terminal. It also adds some project configuration (courtesy of Dustin) to enable F5-debugging of the unpackaged application.

Cherry picked from !3070901

Related work items: #20719950