Merged PR 3083586: Enable emoji to be rendered in the DWrite renderer
This primarily changes the interface used by renderer engines to take a special structure filled with clusters of text information and how many columns they consume in order to allow the DirectWrite renderer to calculate the appropriate spacing for glyphs that hit the font fallback and still need to fit in a certain count of columns.

I've updated all the existing renderers to accept this new format, changed all the function signature interfaces, and so forth.

I've confirmed the current code works for the `foo.txt` sample included that has standard text, emoji, and CJK inline.

Cherry picked from !3061475

Related work items: #20641965, #20641966, #20641967